requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
cors==1.0.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
filelock==3.8.2
future==0.18.2
gevent==22.10.2
greenlet==2.0.1
idna==3.4
PySocks==1.7.1
pytz==2022.6
requests==2.28.1
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
urllib3==1.26.13
zope.event==4.5.0
zope.interface==5.5.2
